requirements.txt
Django==1.5.4
South==0.8.2
Werkzeug==0.9.4
django-extensions==1.2.5
djangorestframework==2.3.8
ipython==1.1.0
six==1.4.1
solrpy==0.9.5
wsgiref==0.1.2
